requirements.txt
packaging
mido
python-rtmidi
keyboard
colorama
pyinstaller
customtkinter
pywin32
pillow
requests